fix tile rendering and water tile selection, fix linter warnings

+++ src/map_genreation/config.rs
@@ -1,7 +1,7 @@
 // Sprites
 pub const TILE_W: usize = 16;
 pub const TILE_H: usize = 16;
-pub const SPRITE_SCALE_FACTOR: usize = 3;
+pub const SPRITE_SCALE_FACTOR: f32 = 3.0;
 pub const SPRITE_SHEET_PATH: &str = "zombie_apocalypse_tileset/tileset.png";
 
 // Window
